moved snack interface to interface folder
