+++ src/interfaces/SnackBarToast.ts
@@ -7,3 +7,10 @@ export interface SnackBarToastInterface {
   snackbarStyle?: string
   severity?: 'success' | 'error' | 'warning' | 'info'
 }
+export interface SnackBarToastProps {
+  snackbarDescription?: React.Dispatch<
+    React.SetStateAction<SnackBarToastInterface>
+  >
+  snackBar?: SnackBarToastInterface
+  snackbarStyle?: string
+}
